+++ src/main/java/dolla/Dolla.java
@@ -1,6 +1,6 @@
 package dolla;
 
-import command.Command;
+import dolla.command.Command;
 import dolla.task.Task;
 import dolla.task.TaskList;
 import parser.MainParser;
@@ -9,27 +9,27 @@ import java.util.ArrayList;
 import java.util.Scanner;
 
 /**
- * <h1>duke.Main</h1>
- * duke.Main is a chat-bot styled todo_list manager.
+ * <h1>duke.Dolla</h1>
+ * duke.Dolla is a chat-bot styled todo_list manager.
  *
  * @author  Aik Peng
  * @version 1.0
  * @since   2019-07-26
  */
-public class Main {
+public class Dolla {
 
     private TaskList tasks = new TaskList(new ArrayList<Task>());
     private String mode = "dolla";
 
     /**
-     * Creates an instance of Main using a data loaded from /data/dolla.txt
+     * Creates an instance of Dolla using a data loaded from /data/dolla.txt
      */
-    public Main() {
+    public Dolla() {
         //tasks = new TaskList(Storage.load());
     }
 
     /**
-     * Runs the main program of duke.Main
+     * Runs the main program of duke.Dolla
      * @throws Exception when exceptional condition happens
      */
     public void run() throws Exception {
@@ -54,6 +54,6 @@ public class Main {
     }
 
     public static void main(String[] args) throws Exception { // Exception needs to be handled?
-        new Main().run();
+        new Dolla().run();
     }
 }
