+++ src/main/java/dolla/Reminder.java
@@ -1,9 +1,8 @@
 package dolla;
 
-import command.Command;
+import dolla.command.Command;
 import dolla.task.Task;
 import dolla.task.TaskList;
-import ui.Ui;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
