+++ src/main/java/dolla/command/AddFixDurationCommand.java
@@ -1,6 +1,6 @@
-package command;
+package dolla.command;
 
-import ui.Ui;
+import dolla.Ui;
 import dolla.task.FixDuration;
 import dolla.task.Task;
 import dolla.task.TaskList;
