+++ src/main/java/dolla/command/AddTodoCommand.java
@@ -1,4 +1,4 @@
-package command;
+package dolla.command;
 
 /**
  * AddToDoCommand is a command used to add a todo task into the duke.task.TaskList.
