+++ src/main/java/dolla/command/ErrorCommand.java
@@ -1,4 +1,7 @@
-package command;
+package dolla.command;
+
+import dolla.Ui;
+import dolla.task.TaskList;
 
 /**
  * duke.command.ErrorCommand is a type of command that is generated if a proper command
