+++ src/main/java/dolla/command/FindStringCommand.java
@@ -1,6 +1,6 @@
-package command;
+package dolla.command;
 
-import ui.Ui;
+import dolla.Ui;
 import dolla.task.Task;
 import dolla.task.TaskList;
