+++ src/test/java/DollaTest.java
@@ -2,7 +2,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class MainTest {
+public class DollaTest {
     @Test
     public void dummyTest() {
         assertEquals(2, 2);
